Restructured code and improved documentation.
This PR is possible because of the dynamic providers module abstraction.

Related to #17.
