--- annotation.go
@@ -1,5 +1,0 @@
-package inject
-
-// Annotation is the interface that has to be implemented by all annotations.
-// It is empty, so implementation is trivial.
-type Annotation interface{}
